yt_concatente/requirements.txt
python-dotenv==1.1.0
webvtt-py==0.5.1
yt-dlp==2025.4.30
